+++ src/app/core/interfaces/admin.interface.ts
@@ -1,0 +1,5 @@
+export interface IAdminState {
+  isLoggedIn: boolean;
+  loading: boolean;
+  isUserDataIncorrect: boolean;
+}
