+++ src/app/core/interfaces/landing.interface.ts
@@ -4,7 +4,4 @@ export interface ILandingState {
   user: IUser;
   loading: boolean;
   userExists: boolean;
-  isLoggedIn: boolean;
-  username: string;
-  isUserDataIncorrect: boolean;
 }
