+++ src/app/store/admin.state.ts
@@ -1,0 +1,91 @@
+import { Injectable } from '@angular/core';
+import Backendless from 'backendless';
+import { Action, State, StateContext, Store } from '@ngxs/store';
+import { Router } from '@angular/router';
+
+import {
+  LoginUser,
+  LoginUserError,
+  LoginUserSuccess,
+  UserLoggedIn,
+  UserLoggedInSuccess,
+  UserNotLoggedIn,
+  LogoutUser
+} from './admin.action';
+import { IAdminState } from '../core/interfaces/admin.interface';
+
+@State<IAdminState>({
+  name: 'adminState',
+  defaults: {
+    isLoggedIn: false,
+    loading: false,
+    isUserDataIncorrect: false
+  }
+})
+@Injectable()
+export class AdminState {
+  constructor(private router: Router, private store: Store) {}
+
+  @Action(LoginUser)
+  loginUser(ctx: StateContext<IAdminState>, { user }: LoginUser): void {
+    ctx.patchState({ loading: true });
+
+    Backendless.UserService.login(user.email, user.password, true)
+      .then((loggedInUser) => {
+        console.log(loggedInUser);
+
+        this.store.dispatch(new LoginUserSuccess(loggedInUser));
+      })
+      .catch((error) => {
+        console.log(error);
+
+        this.store.dispatch(new LoginUserError());
+      });
+  }
+
+  @Action(LoginUserSuccess)
+  loginUserSuccess(ctx: StateContext<IAdminState>): void {
+    ctx.patchState({
+      loading: false,
+      isLoggedIn: true,
+      isUserDataIncorrect: false
+    });
+  }
+
+  @Action(LoginUserError)
+  loginUserError(ctx: StateContext<IAdminState>): void {
+    ctx.patchState({ loading: false, isUserDataIncorrect: true });
+  }
+
+  @Action(UserLoggedIn)
+  userLoggedIn(): void {
+    Backendless.UserService.getCurrentUser()
+      .then((user) => {
+        this.store.dispatch(new UserLoggedInSuccess(user));
+      })
+      .catch(() => {
+        this.store.dispatch(new UserNotLoggedIn());
+      });
+  }
+
+  @Action(UserLoggedInSuccess)
+  userLoggedInSuccess(ctx: StateContext<IAdminState>): void {
+    ctx.patchState({ isLoggedIn: true });
+  }
+
+  @Action(UserNotLoggedIn)
+  userNotLoggedIn(ctx: StateContext<IAdminState>): void {
+    ctx.patchState({ isLoggedIn: false });
+  }
+
+  @Action(LogoutUser)
+  logOut(ctx: StateContext<IAdminState>): void {
+    Backendless.UserService.logout()
+      .then(() => {
+        ctx.patchState({ isLoggedIn: false });
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  }
+}
