+++ src/app/store/landing.state.ts
@@ -4,20 +4,13 @@ import { Action, State, StateContext, Store } from '@ngxs/store';
 import { Router } from '@angular/router';
 
 import {
-  LoginUser,
-  LoginUserError,
-  LoginUserSuccess,
-  LogoutUser,
   RegisterUser,
   RegisterUserError,
   RegisterUserSuccess,
   SendMessage,
   SendMessageError,
   SendMessageSuccess,
-  UserExists,
-  UserLoggedIn,
-  UserLoggedInSuccess,
-  UserNotLoggedIn
+  UserExists
 } from './landing.action';
 import { IRegisterRequest } from '../core/interfaces/register.interface';
 import { ILandingState } from '../core/interfaces/landing.interface';
@@ -27,10 +20,7 @@ import { ILandingState } from '../core/interfaces/landing.interface';
   defaults: {
     user: null,
     loading: false,
-    userExists: false,
-    isLoggedIn: false,
-    username: '',
-    isUserDataIncorrect: false
+    userExists: false
   }
 })
 @Injectable()
@@ -121,71 +111,5 @@ export class LandingState {
   @Action(SendMessageError)
   sendMessageError(ctx: StateContext<ILandingState>): void {
     ctx.patchState({ loading: false });
-  }
-
-  @Action(LoginUser)
-  loginUser(ctx: StateContext<ILandingState>, { user }: LoginUser): void {
-    ctx.patchState({ loading: true });
-
-    Backendless.UserService.login(user.email, user.password, true)
-      .then((loggedInUser) => {
-        console.log(loggedInUser);
-
-        this.store.dispatch(new LoginUserSuccess(loggedInUser));
-      })
-      .catch((error) => {
-        console.log(error);
-
-        this.store.dispatch(new LoginUserError());
-      });
-  }
-
-  @Action(LoginUserSuccess)
-  loginUserSuccess(ctx: StateContext<ILandingState>): void {
-    ctx.patchState({
-      loading: false,
-      isLoggedIn: true,
-      isUserDataIncorrect: false
-    });
-  }
-
-  @Action(LoginUserError)
-  loginUserError(ctx: StateContext<ILandingState>): void {
-    ctx.patchState({ loading: false, isUserDataIncorrect: true });
-  }
-
-  @Action(UserLoggedIn)
-  userLoggedIn(): void {
-    Backendless.UserService.getCurrentUser()
-      .then((user) => {
-        this.store.dispatch(new UserLoggedInSuccess(user));
-      })
-      .catch(() => {
-        this.store.dispatch(new UserNotLoggedIn());
-      });
-  }
-
-  @Action(UserLoggedInSuccess)
-  userLoggedInSuccess(
-    ctx: StateContext<ILandingState>,
-    { user }: UserLoggedInSuccess
-  ): void {
-    ctx.patchState({ isLoggedIn: true, username: user.username });
-  }
-
-  @Action(UserNotLoggedIn)
-  userNotLoggedIn(ctx: StateContext<ILandingState>): void {
-    ctx.patchState({ isLoggedIn: false, username: '' });
-  }
-
-  @Action(LogoutUser)
-  logOut(ctx: StateContext<ILandingState>): void {
-    Backendless.UserService.logout()
-      .then(() => {
-        ctx.patchState({ isLoggedIn: false });
-      })
-      .catch((error) => {
-        console.log(error);
-      });
   }
 }
